ECL exp

+++ GSL/utils/score.py
@@ -21,7 +21,7 @@ def get_score(target, prediction, scaler=None):
     return score
 
 def MAPE(v, v_, axis=None):
-    mape = (np.abs(v_ - v) / (np.abs(v)+1e-5)).astype(np.float64)
+    mape = (np.abs(v_ - v) / (np.abs(v)+1e-5))
     mape = np.where(mape > 5, 5, mape)
     return np.mean(mape, axis)
 
@@ -33,7 +33,7 @@ def RMSE(v, v_, axis=None):
     :param axis: axis to do calculation.
     :return: int, RMSE averages on all elements of input.
     '''
-    return np.sqrt(np.mean((v_ - v) ** 2, axis)+1e-5).astype(np.float64)
+    return np.sqrt(np.mean((v_ - v) ** 2, axis)+1e-5)
 
 
 def MAE(v, v_, axis=None):
@@ -44,4 +44,4 @@ def MAE(v, v_, axis=None):
     :param axis: axis to do calculation.
     :return: int, MAE averages on all elements of input.
     '''
-    return np.mean(np.abs(v_ - v), axis).astype(np.float64)
+    return np.mean(np.abs(v_ - v), axis)
